feat(ChildCommmentActivity): 구성 변경 시 방금 들어왔다는 플래그를 false로 설정
